[Chore] .storybook/preview에 global style과 theme 추가
